updater will delete itself

+++ extras/menus/xMenu/include/text.h
@@ -24,7 +24,7 @@ class TextAnim{
 
     private:
 
-        float scroll;
+        int scroll;
         string title;
         string subtitle;
